test: run get function in a seperate process
