+++ package.json
@@ -44,7 +44,7 @@
   "scripts": {
     "lint": "eslint src/get.js",
     "docs": "jsdoc -d docs ./src/get.js ./src/run.js ./src/bld.js",
-    "test": "node test/index.js",
+    "test": "node test/get.pre.js && node --test test/get.test.js",
     "demo": "cd test/fixture && node demo.js"
   },
   "devDependencies": {
