--- test/index.js
@@ -1,7 +1,0 @@
-// import * as addonTests from "./addon.js";
-import * as getTests from "./get.js";
-// import * as modeTests from "./mode.js";
-
-// addonTests;
-getTests;
-// modeTests;
